pyproject.toml
[tool.poetry]
name = "k8-pod-notifier"
version = "0.1.0"
description = ""
authors = ["Muhammad Jamali <muhammadjamali003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kubernetes = "^29.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
